requirements.txt
fastapi==0.121.3
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.25
Alembic==1.13.1
python-multipart==0.0.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.6.3
pydantic-settings==2.2.1
cryptography==42.0.5
httpx==0.27.0
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.2.0
python-dotenv==1.0.1
itsdangerous==2.1.2
orjson==3.9.15
loguru==0.7.2
